+++ sqlalchemy_fake_model/Enum/ModelColumnTypesEnum.py
@@ -1,8 +1,26 @@
 from enum import Enum
 
-from sqlalchemy import Boolean, Date, DateTime, Float, Integer, String, Text
+from sqlalchemy import (
+    Boolean,
+    Date,
+    DateTime,
+    Float,
+    Integer,
+    Interval,
+    LargeBinary,
+    String,
+    Text,
+    Time,
+)
 from sqlalchemy import Enum as SQLAlchemyEnum
+from sqlalchemy.dialects.postgresql import JSON, JSONB, UUID
 from sqlalchemy.orm import relationship
+from sqlalchemy.types import DECIMAL
+
+try:
+    from sqlalchemy.dialects.postgresql import UUID as PostgreSQLUUID
+except ImportError:
+    PostgreSQLUUID = None
 
 
 class ModelColumnTypesEnum(Enum):
@@ -21,6 +39,20 @@ class ModelColumnTypesEnum(Enum):
     DATETIME = DateTime
 
     DATE = Date
+
+    TIME = Time
+
+    INTERVAL = Interval
+
+    DECIMAL = DECIMAL
+
+    LARGEBINARY = LargeBinary
+
+    UUID = UUID if PostgreSQLUUID else String
+
+    JSON = JSON
+
+    JSONB = JSONB
 
     ENUM = SQLAlchemyEnum