+++ sqlalchemy_fake_model/__init__.py
@@ -1,1 +1,4 @@
 from .ModelFaker import ModelFaker
+from .SmartFieldDetector import SmartFieldDetector
+
+__all__ = ["ModelFaker", "SmartFieldDetector"]